+++ src/components/DailyReport/index.tsx
@@ -87,9 +87,6 @@ const DailyReportShare: React.FC<DailyReportShareProps> = (props: DailyReportSha
 };
 
 
-
-export type { DailyReportProps } from '../../types/components';
-
 export const DailyReport: React.FC<DailyReportProps> = (props: DailyReportProps) => {
   const [dailyReportType, setDailyReportType] = useState<DailyReportType>(DailyReportType.HTML);
