+++ src/components/EsaSubmitForm/index.tsx
@@ -47,8 +47,6 @@ export const submitTextToEsa = (
   return submitTextToEsaApi(category, tags, title, text);
 };
 
-export type { EsaSubmitFormProps } from '../../types/components';
-
 export const EsaSubmitForm: React.FC<EsaSubmitFormProps> = (props: EsaSubmitFormProps) => {
   const [sending, setSending] = useState(false);
   const [category, setCategory] = useState<string>(props.category);
