+++ src/components/Footer/index.tsx
@@ -6,9 +6,6 @@ import { config } from '../../config';
 import { GoogleUser } from '../../types';
 import { FooterProps } from '../../types/components';
 
-
-export type { FooterProps } from '../../types/components';
-
 export const Footer: React.FC<FooterProps> = (props: FooterProps) => {
   const hr = (
     <hr style={{
